Move SQLMesh Airflow API into a separate module
